refactor: 클래스 위치 변경 (게임과 게임 결과 분리)
